My fix got yoinked
??

+++ resources/[esx]/es_extended/client/main.lua
@@ -351,8 +351,10 @@ end)
 
 RegisterNetEvent('esx:removePickup')
 AddEventHandler('esx:removePickup', function(id)
-	ESX.Game.DeleteObject(Pickups[id].obj)
-	Pickups[id] = nil
+  if (id ~= nil) then
+    ESX.Game.DeleteObject(Pickups[id].obj)
+	  Pickups[id] = nil
+  end
 end)
 
 RegisterNetEvent('esx:pickupWeapon')
@@ -388,14 +390,11 @@ AddEventHandler('esx:deleteVehicle', function()
 local entity
     local veh = GetVehiclePedIsIn(GetPlayerPed(PlayerId()), false)
     if veh == -1 then
-      local playerPed = PlayerPedId()
-      local vehicle   = ESX.Game.GetVehicleInDirection()
-      entity = vehicle
-      carModel = GetEntityModel(entity)
-      carName = GetDisplayNameFromVehicleModel(carModel)
+      local loc = GetEntityCoords(GetPlayerPed(PlayerPedId()))
+      local entity   = GetClosestVehicle(loc.x, loc.y, loc.z, 2, 0, 71)
       NetworkRequestControlOfEntity(entity)
     else
-      NetworkRequestControlOfEntity(GetEntityModel(veh))
+      NetworkRequestControlOfEntity(veh)
       entity = veh
     end
